+++ wam/settings.py
@@ -209,3 +209,8 @@ for app in WAM_APPS:
 WAM_EXCHANGE_ACCOUNT = config['WAM']['WAM_EXCHANGE_ACCOUNT']
 WAM_EXCHANGE_EMAIL = config['WAM']['WAM_EXCHANGE_EMAIL']
 WAM_EXCHANGE_PW = config['WAM']['WAM_EXCHANGE_PW']
+
+# Extensions for markdownx
+MARKDOWNX_MARKDOWN_EXTENSIONS = [
+    'markdown.extensions.extra'
+]