+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Here is a template for new release sections
 
 ### Added
 - I18N functionality via locale middleware enabled #76 #80
+- enable support for markdownx extensions #83
 
 ### Changed